add parallel cross_entropy unit test to auto_parallel
